+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This file is manually updated. Please file an issue if something is missing.
 
 ## v4.8.0, v3.8.0
 
+- **Feature**: Support for Live Photos
+  - You need to run `occ memories:index --clear` to reindex live photos
+  - Only JPEG (iOS with MOV, Google, Samsung) is supported. HEIC is not supported.
 - **Feature**: Timeline path now scans recursively for mounted volumes / shares inside it
 - **Feature**: Multiple timeline paths can be specified ([#178](https://github.com/pulsejet/memories/issues/178))
 - Support for server-side encrypted storage ([#99](https://github.com/pulsejet/memories/issues/99))